modulus_feature_added

+++ Embedded_Linux/EmbeddedLinux_Task01/Src/modulus.c
@@ -3,5 +3,5 @@
 
 double Remainder(double dividend ,double divisor)
 {
-	return ( dividend % divisor );
+	return ( (long long)dividend % (long long)divisor );
 }
